Flutter (Android)
As a junior developer at Global New Wave in Myanmar, the task is to create a prototype app that encourages users to "save products for later" instead of making immediate purchases. The app allows users to revisit their saved products and access the product homepage via a link. Users can also remove items they no longer want. The goal is to design a user-friendly and efficient mobile app that meets customer needs.